--- src/components/cards/main.module.sass
@@ -1,15 +1,0 @@
-.link-card
-  padding: 1em
-  margin: 1em 0 1em
-  display: block
-  text-decoration: none
-  color: unset
-  &:hover
-    text-decoration: none
-    background-color: var(--accent-secondary-hover-color)
-  :first-child
-    margin-top: 0
-  :last-child
-    margin-bottom: 0
-
-
