--- src/components/icon.module.sass
@@ -1,21 +1,0 @@
-.icon-container
-  display: inline-block
-  overflow: hidden
-  align-items: center
-  justify-content: center
-  border: 1px solid #000
-  flex-shrink: 0
-  cursor: pointer
-
-.page-title
-  align-items: center
-  .icon-container
-    margin-right: 0.4em
-    margin-bottom: 0
-    transform: translateY(0.05em)
-  .subtitle
-    margin-left: 1em
-    font-weight: 300
-  .site-name
-    font-weight: 300
-    color: unset
