--- src/components/map-navbar/main.module.sass
@@ -1,22 +1,0 @@
-.map-title
-  // Don't allow the title to wrap
-  white-space: nowrap
-  overflow-x: hidden
-  text-overflow: ellipsis
-
-.map-navbar
-  transition: min-width 0.5s
-  :global(.bp5-navbar)
-    &:hover
-      min-width: fit-content
-      .map-title
-        // Allow the title to wrap on hover
-        overflow-x: visible
-
-.map-minimal-navbar
-  display: flex
-  flex-direction: row
-  &>.spacer
-    flex-grow: 1
-  .map-navbar
-    min-width: 0
